web/cryptkeeper/requirements.txt
Django==2.0.10
biopython==1.71
plotly==3.0.0
